requirements.txt
gunicorn>=20.0.4
plotly>=4.5.0
dash==1.4.1
numpy>=1.18
requests>=2.22
pandas>=1.0.1
Pillow>=8.3.0
matplotlib>=3.4.0
